feat: Add Dockerfile for containerizing the application
This commit adds a Dockerfile to the repository, which allows the application to be containerized using Docker. The Dockerfile specifies the base Python image, sets the working directory, installs build dependencies, copies the requirements file, installs the dependencies, and copies the source code. It also exposes the port on which the FastAPI application will run and sets the command to start the application. This addition enhances the deployment and scalability of the application by providing a standardized and reproducible environment.